Fixed: Both TMath#periodic24 and the test for it were broken because it was not casting the result of the integer function to a float before dividing by the range of an integer and the test was calling the wrong function!

+++ src/togos/networkrts/util/TMath.java
@@ -29,6 +29,6 @@ public final class TMath
 	 * where precision doesn't matter and is ~2 orders of magnitude faster.
 	 */
 	public static final float periodic24( int input ) {
-		return periodic24_32( input ) / (65536*32767);
+		return (float)periodic24_32( input ) / (65536*32767);
 	}
 }
